+++ test/blocks/change-bg/mocks/body.html
@@ -1,1 +1,5 @@
-<div class="change-bg" daa-lh="b1|change-bg|default|default"><ft-changebackgroundmarquee></ft-changebackgroundmarquee></div>
+<div class="interactive-marquee" data-block="">
+    <div>
+      <div>https://main--cc--suhjainadobe.hlx.live/drafts/suhjain/interactive-changebg.json</div>
+    </div>
+  </div>
